feat(playground): integrate playground into navigation
