+++ docs/content/2.get-started.md
@@ -5,7 +5,6 @@ icon: heroicons-outline:lightning-bolt
 layout: blogpost
 ---
 
-
 ## Play online
 
 You can start playing with Nuxt Content in your browser using our online sandboxes:
@@ -20,6 +19,7 @@ You can start playing with Nuxt Content in your browser using our online sandbox
   ::details
     :summary[Before getting started, please **make sure to have installed the recommended setup**:]
     * Follow the [Nuxt 3 Prerequisites](https://v3.nuxtjs.org/getting-started/quick-start#prerequisites)
+    
     * Download the [VS Code **MDC Extension**](https://marketplace.visualstudio.com/items?itemName=Nuxt.mdc)
   ::
 ::
