+++ docs/content/playground.md
@@ -1,0 +1,5 @@
+---
+layout: fluid
+---
+
+::playground
